allowing user defined path to readme and a MKDocs config file
