+++ render-and-deploy-docs/Dockerfile
@@ -3,7 +3,6 @@ FROM python:3.8.1
 RUN apt-get update && apt-get install -y git libssl-dev build-essential
 
 RUN mkdir /src
-WORKDIR /src
 
 COPY ./ /src/
