+++ render-and-deploy-docs/action.yml
@@ -1,7 +1,20 @@
 ---
 name: 'Render and deploy docs'
 description: 'Render pipeline documentation and deploy to GH-Pages'
+inputs:
+  readme:
+    description: 'Relative path to the README file.'
+    required: false
+    default: 'README.md'
+  mkdocs_yaml:
+    description: 'Relative path to the MKDocs config yaml.'
+    required: false
+    default: 'None'
 
 runs:
   using: 'docker'
   image: 'Dockerfile'
+  entrypoint: '/src/action.sh'
+  args:
+    - ${{ inputs.mkdocs_yaml }}
+    - ${{ inputs.readme }}
